Update patch set 3
Patch Set 3:

Source built shouldn't matter (its out of the if/else block). If you are just building the Gello target vs the whole CM system, that would  explain the double up. But lets not focus on that as nightlies use the maven artifact; the partition question is more interesting. 

685768704 is indeed the right partition size for system 

https://android.googlesource.com/device/samsung/manta/+/master/BoardConfig.mk#55

Was there a manta with a larger system partition sold or is your's always been resized?

Patch-set: 3