Update patch set 3
Patch Set 3:

Scott,

keep in mind that df shows the size of the filesystem, not the
size of the partition or block device.

If you go into a root shell (e.g. via adb), like this:

 host$ adb shell
 shell@manta:/ $ su
 root@manta:/ # dd if=/dev/block/platform/dw_mmc.0/by-name/system | wc -c

what's the output?  In my case it's the aforementioned 838860800.

Also, if you go into TWRP recovery, enter a shell and perform the
following actions:

  ~ # umount /system  (only if mounted, not by default)
  ~ # e2fsck /dev/block/platform/dw_mmc.0/by-name/system
  ~ # resize2fs /dev/block/platform/dw_mmc.0/by-name/system
  ~ # mount /system
  ~ # df

what's the output of df afterwards?


Thanks,
Corinna

Patch-set: 3